Move remaining invoke_training.training.shared files to training2.
