+++ src/invoke_training/training/finetune_lora/finetune_lora_sdxl.py
@@ -22,19 +22,19 @@ from invoke_training.core.lora.injection.stable_diffusion import (
 )
 from invoke_training.training.config.finetune_lora_config import FinetuneLoRASDXLConfig
 from invoke_training.training.finetune_lora.tokenize_captions import tokenize_captions
-from invoke_training.training.shared.accelerator_utils import (
+from invoke_training.training2.shared.accelerator.accelerator_utils import (
     get_mixed_precision_dtype,
     initialize_accelerator,
     initialize_logging,
 )
-from invoke_training.training.shared.checkpoint_tracker import CheckpointTracker
-from invoke_training.training.shared.lora_checkpoint_utils import save_lora_checkpoint
-from invoke_training.training.shared.model_loading_utils import PipelineVersionEnum, load_pipeline
-from invoke_training.training.shared.optimizer_utils import initialize_optimizer
+from invoke_training.training2.shared.checkpoints.checkpoint_tracker import CheckpointTracker
 from invoke_training.training2.shared.data.data_loaders.image_caption_sdxl_dataloader import (
     build_image_caption_sdxl_dataloader,
 )
 from invoke_training.training2.shared.data.transforms.tensor_disk_cache import TensorDiskCache
+from invoke_training.training2.shared.optimizer.optimizer_utils import initialize_optimizer
+from invoke_training.training2.shared.stable_diffusion.lora_checkpoint_utils import save_lora_checkpoint
+from invoke_training.training2.shared.stable_diffusion.model_loading_utils import PipelineVersionEnum, load_pipeline
 
 
 def _import_model_class_for_model(pretrained_model_name_or_path: str, subfolder: str = "", revision: str = "main"):
