+++ src/invoke_training/training/textual_inversion/textual_inversion_sd.py
@@ -20,18 +20,18 @@ from invoke_training.training.finetune_lora.finetune_lora_sd import (
     generate_validation_images,
     train_forward,
 )
-from invoke_training.training.shared.accelerator_utils import (
+from invoke_training.training2.shared.accelerator.accelerator_utils import (
     get_mixed_precision_dtype,
     initialize_accelerator,
     initialize_logging,
 )
-from invoke_training.training.shared.checkpoint_tracker import CheckpointTracker
-from invoke_training.training.shared.model_loading_utils import PipelineVersionEnum, load_pipeline
-from invoke_training.training.shared.optimizer_utils import initialize_optimizer
-from invoke_training.training.shared.serialization import load_state_dict, save_state_dict
+from invoke_training.training2.shared.checkpoints.checkpoint_tracker import CheckpointTracker
+from invoke_training.training2.shared.checkpoints.serialization import load_state_dict, save_state_dict
 from invoke_training.training2.shared.data.data_loaders.textual_inversion_sd_dataloader import (
     build_textual_inversion_sd_dataloader,
 )
+from invoke_training.training2.shared.optimizer.optimizer_utils import initialize_optimizer
+from invoke_training.training2.shared.stable_diffusion.model_loading_utils import PipelineVersionEnum, load_pipeline
 
 
 def load_models(
