+++ src/invoke_training/training2/shared/checkpoints/serialization.py
