+++ src/invoke_training/training2/shared/stable_diffusion/base_model_version.py
