+++ src/invoke_training/training2/shared/stable_diffusion/model_loading_utils.py
