+++ tests/invoke_training/training/shared/checkpoints/test_checkpoint_tracker.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from invoke_training.training.shared.checkpoint_tracker import CheckpointTracker
+from invoke_training.training2.shared.checkpoints.checkpoint_tracker import CheckpointTracker
 
 
 def test_checkpoint_tracker_get_path_file():
